DO_NOT_MERGE: experiments in compiling on raspberry pi
There are quite a few crude hacks in here.

+++ src/tests/vnet/session/tcp_echo.c
@@ -1277,7 +1277,7 @@ main (int argc, char **argv)
 
   clib_time_init (&em->clib_time);
   init_error_string_table (em);
-  svm_fifo_segment_main_init (0x200000000ULL, 20);
+  svm_fifo_segment_main_init ((ssize_t)0x200000000ULL, 20);
   unformat_init_command_line (a, argv);
 
   while (unformat_check_input (a) != UNFORMAT_END_OF_INPUT)
